renamed scIB to scib
