+++ scib/exceptions.py
