+++ scib/metrics/kbet.py
@@ -7,7 +7,7 @@ import rpy2.rinterface_lib.callbacks
 import anndata2ri
 import scanpy as sc
 
-from scIB.utils import checkAdata, checkBatch
+from ..utils import checkAdata, checkBatch
 from .utils import diffusion_conn, diffusion_nn, NeighborsError
 
 rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR)  # Ignore R warning messages
