+++ scib/metrics/silhouette.py
