+++ scib/metrics/trajectory.py
