+++ scib/preprocessing.py
@@ -7,18 +7,18 @@ from scipy import sparse
 # rpy2 for running R code
 import rpy2.rinterface_lib.callbacks
 import logging
-rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR) # Ignore R warning messages
+
+rpy2.rinterface_lib.callbacks.logger.setLevel(logging.ERROR)  # Ignore R warning messages
 import rpy2.robjects as ro
 import anndata2ri
 
 # access to other methods of this module
-from scIB.utils import *
+from .utils import *
 
 
 def summarize_counts(adata, count_matrix=None, mt_gene_regex='^MT-'):
-    
     checkAdata(adata)
-    
+
     if count_matrix is None:
         count_matrix = adata.X
     adata.obs['n_counts'] = count_matrix.sum(1)
@@ -34,32 +34,32 @@ def summarize_counts(adata, count_matrix=None, mt_gene_regex='^MT-'):
         if sparse.issparse(adata.X):
             mt_sum = mt_sum.A1
             total_sum = total_sum.A1
-        adata.obs['percent_mito'] =  mt_sum / total_sum 
-        
-        #mt_gene_mask = [gene.startswith('mt-') for gene in adata.var_names]
-        #mt_count = count_matrix[:, mt_gene_mask].sum(1)
-        #if mt_count.ndim > 1:
-        #    mt_count = np.squeeze(np.asarray(mt_count))
-        #adata.obs['mt_frac'] = mt_count/adata.obs['n_counts']
+        adata.obs['percent_mito'] = mt_sum / total_sum
 
-### Quality Control        
+        # mt_gene_mask = [gene.startswith('mt-') for gene in adata.var_names]
+        # mt_count = count_matrix[:, mt_gene_mask].sum(1)
+        # if mt_count.ndim > 1:
+        #    mt_count = np.squeeze(np.asarray(mt_count))
+        # adata.obs['mt_frac'] = mt_count/adata.obs['n_counts']
+
+
+### Quality Control
 def plot_QC(adata, color=None, bins=60, legend_loc='right margin', histogram=True,
-            gene_threshold=(0,np.inf), 
-            gene_filter_threshold=(0,np.inf),
-            count_threshold=(0,np.inf), 
-            count_filter_threshold=(0,np.inf)):
-    
+            gene_threshold=(0, np.inf),
+            gene_filter_threshold=(0, np.inf),
+            count_threshold=(0, np.inf),
+            count_filter_threshold=(0, np.inf)):
     if count_filter_threshold == (0, np.inf):
         count_filter_threshold = count_threshold
     if gene_filter_threshold == (0, np.inf):
         gene_filter_threshold = gene_threshold
-    
+
     # 2D scatter plot
     plot_scatter(adata, color=color, title=color,
-                 gene_threshold=gene_filter_threshold[0], 
+                 gene_threshold=gene_filter_threshold[0],
                  count_threshold=count_filter_threshold[0],
                  legend_loc=legend_loc)
-    
+
     if not histogram:
         return
 
@@ -67,84 +67,84 @@ def plot_QC(adata, color=None, bins=60, legend_loc='right margin', histogram=Tru
         print(f"Counts Threshold: {count_filter_threshold}")
         # count filtering
         plot_count_filter(adata, obs_col='n_counts', bins=bins,
-                          lower = count_threshold[0],
-                          filter_lower = count_filter_threshold[0],
-                          upper = count_threshold[1],
-                          filter_upper = count_filter_threshold[1])
-    
+                          lower=count_threshold[0],
+                          filter_lower=count_filter_threshold[0],
+                          upper=count_threshold[1],
+                          filter_upper=count_filter_threshold[1])
+
     if gene_filter_threshold != (0, np.inf):
         print(f"Gene Threshold: {gene_filter_threshold}")
         # gene filtering
         plot_count_filter(adata, obs_col='n_genes', bins=bins,
-                          lower = gene_threshold[0],
-                          filter_lower = gene_filter_threshold[0],
-                          upper = gene_threshold[1],
-                          filter_upper = gene_filter_threshold[1])
+                          lower=gene_threshold[0],
+                          filter_lower=gene_filter_threshold[0],
+                          upper=gene_threshold[1],
+                          filter_upper=gene_filter_threshold[1])
+
 
 def plot_scatter(adata, count_threshold=0, gene_threshold=0,
                  color=None, title='', lab_size=15, tick_size=11, legend_loc='right margin',
                  palette=None):
-    
     checkAdata(adata)
     if color:
         checkBatch(color, adata.obs)
-    
+
     ax = sc.pl.scatter(adata, 'n_counts', 'n_genes', color=color, show=False,
                        legend_fontweight=50, legend_loc=legend_loc, palette=palette)
     ax.set_title(title, fontsize=lab_size)
-    ax.set_xlabel("Count depth",fontsize=lab_size)
-    ax.set_ylabel("Number of genes",fontsize=lab_size)
+    ax.set_xlabel("Count depth", fontsize=lab_size)
+    ax.set_ylabel("Number of genes", fontsize=lab_size)
     ax.tick_params(labelsize=tick_size)
-    
+
     if gene_threshold > 0:
-        ax.axhline(gene_threshold, 0,1, color='red')
+        ax.axhline(gene_threshold, 0, 1, color='red')
     if count_threshold > 0:
-        ax.axvline(count_threshold, 0,1, color='red')
-    
+        ax.axvline(count_threshold, 0, 1, color='red')
+
     fig = plt.gcf()
     cbar_ax = fig.axes[-1]
     cbar_ax.tick_params(labelsize=tick_size)
     f1 = ax.get_figure()
     plt.show()
 
+
 def plot_count_filter(adata, obs_col='n_counts', bins=60, lower=0, upper=np.inf, filter_lower=0, filter_upper=np.inf):
-    
     plot_data = adata.obs[obs_col]
-    
+
     sns.distplot(plot_data, kde=False, bins=bins)
-    
+
     if lower > 0:
-        plt.axvline(lower, linestyle = '--', color = 'g')
+        plt.axvline(lower, linestyle='--', color='g')
     if filter_lower > 0:
-        plt.axvline(filter_lower, linestyle = '-', color = 'r')
+        plt.axvline(filter_lower, linestyle='-', color='r')
     if not np.isinf(upper):
-        plt.axvline(upper, linestyle = '--', color = 'g')
+        plt.axvline(upper, linestyle='--', color='g')
     if not np.isinf(upper):
-        plt.axvline(filter_upper, linestyle = '-', color = 'r')
-    plt.ticklabel_format(style='sci', axis='x', scilimits=(0,2))
+        plt.axvline(filter_upper, linestyle='-', color='r')
+    plt.ticklabel_format(style='sci', axis='x', scilimits=(0, 2))
     plt.show()
-    
+
     # determine lower bound of total, look at points below lower bound
     if filter_lower > 0:
         print(f"lower threshold: {filter_lower}")
         sns.distplot(plot_data[plot_data < lower], kde=False, bins=bins)
-        plt.axvline(filter_lower, linestyle = '-', color = 'r')
-        plt.axvline(lower, linestyle = '--', color = 'g')
-        plt.ticklabel_format(style='sci', axis='x', scilimits=(0,2))
+        plt.axvline(filter_lower, linestyle='-', color='r')
+        plt.axvline(lower, linestyle='--', color='g')
+        plt.ticklabel_format(style='sci', axis='x', scilimits=(0, 2))
         plt.show()
 
     # determine upper bound of total
     if not np.isinf(filter_upper) and not np.isinf(upper):
         print(f"upper threshold: {filter_upper}")
         sns.distplot(plot_data[plot_data > upper], kde=False, bins=bins)
-        plt.axvline(filter_upper, linestyle = '-', color = 'r')
-        plt.axvline(upper, linestyle = '--', color = 'g')
-        plt.ticklabel_format(style='sci', axis='x', scilimits=(0,2))
+        plt.axvline(filter_upper, linestyle='-', color='r')
+        plt.axvline(upper, linestyle='--', color='g')
+        plt.ticklabel_format(style='sci', axis='x', scilimits=(0, 2))
         plt.show()
 
+
 ### Normalisation
-def normalize(adata, min_mean = 0.1, log=True, precluster=True, sparsify=True):
-    
+def normalize(adata, min_mean=0.1, log=True, precluster=True, sparsify=True):
     checkAdata(adata)
 
     # Check for 0 count cells
@@ -159,22 +159,22 @@ def normalize(adata, min_mean = 0.1, log=True, precluster=True, sparsify=True):
 
     if sparsify:
         # massive speedup when working with sparse matrix
-        if not sparse.issparse(adata.X): # quick fix: HVG doesn't work on dense matrix
+        if not sparse.issparse(adata.X):  # quick fix: HVG doesn't work on dense matrix
             adata.X = sparse.csr_matrix(adata.X)
 
     anndata2ri.activate()
     ro.r('library("scran")')
-    
+
     # keep raw counts
     adata.layers["counts"] = adata.X.copy()
 
-    is_sparse=False
+    is_sparse = False
     X = adata.X.T
     # convert to CSC if possible. See https://github.com/MarioniLab/scran/issues/70
     if sparse.issparse(X):
         is_sparse = True
-        
-        if X.nnz > 2**31-1:
+
+        if X.nnz > 2 ** 31 - 1:
             X = X.tocoo()
         else:
             X = X.tocsc()
@@ -200,10 +200,10 @@ def normalize(adata, min_mean = 0.1, log=True, precluster=True, sparsify=True):
     else:
         size_factors = ro.r('sizeFactors(computeSumFactors(SingleCellExperiment('
                             f'list(counts=data_mat)), min.mean = {min_mean}))')
-        
+
     # modify adata
     adata.obs['size_factors'] = size_factors
-    adata.X /= adata.obs['size_factors'].values[:,None]
+    adata.X /= adata.obs['size_factors'].values[:, None]
     if log:
         print("Note! Performing log1p-transformation after normalization.")
         sc.pp.log1p(adata)
@@ -214,7 +214,7 @@ def normalize(adata, min_mean = 0.1, log=True, precluster=True, sparsify=True):
         # convert to sparse, bc operation always converts to dense
         adata.X = sparse.csr_matrix(adata.X)
 
-    adata.raw = adata # Store the full data set in 'raw' as log-normalised data for statistical testing
+    adata.raw = adata  # Store the full data set in 'raw' as log-normalised data for statistical testing
 
     # Free memory in R
     ro.r('rm(list=ls())')
@@ -257,12 +257,13 @@ def scale_batch(adata, batch):
 
     del tmp
     del adata_copy
-    
+
     return adata_scaled
 
-    
-def hvg_intersect(adata, batch, target_genes=2000, flavor='cell_ranger', n_bins=20, adataOut=False, n_stop=8000, min_genes=500, step_size=1000):
-### Feature Selection
+
+def hvg_intersect(adata, batch, target_genes=2000, flavor='cell_ranger', n_bins=20, adataOut=False, n_stop=8000,
+                  min_genes=500, step_size=1000):
+    ### Feature Selection
     """
     params:
         adata:
@@ -273,48 +274,47 @@ def hvg_intersect(adata, batch, target_genes=2000, flavor='cell_ranger', n_bins=
     return:
         list of highly variable genes less or equal to `target_genes`
     """
-    
+
     checkAdata(adata)
     checkBatch(batch, adata.obs)
-    
+
     intersect = None
     enough = False
     n_hvg = target_genes
-    
-    split = splitBatches(adata, batch) 
+
+    split = splitBatches(adata, batch)
     hvg_res = []
 
     for i in split:
-        sc.pp.filter_genes(i, min_cells=1) # remove genes unexpressed (otherwise hvg might break)
+        sc.pp.filter_genes(i, min_cells=1)  # remove genes unexpressed (otherwise hvg might break)
         hvg_res.append(sc.pp.highly_variable_genes(i, flavor='cell_ranger', n_top_genes=n_hvg, inplace=False))
 
     while not enough:
         genes = []
 
         for i in range(len(split)):
-
             dispersion_norm = hvg_res[i]['dispersions_norm']
             dispersion_norm = dispersion_norm[~np.isnan(dispersion_norm)]
             dispersion_norm[::-1].sort()
-            disp_cut_off = dispersion_norm[n_hvg-1]
+            disp_cut_off = dispersion_norm[n_hvg - 1]
             gene_subset = np.nan_to_num(hvg_res[i]['dispersions_norm']) >= disp_cut_off
 
             genes.append(set(split[i].var[gene_subset].index))
 
         intersect = genes[0].intersection(*genes[1:])
-        if len(intersect)>=target_genes:
-            enough=True
+        if len(intersect) >= target_genes:
+            enough = True
         else:
-            if n_hvg>n_stop:
+            if n_hvg > n_stop:
                 if len(intersect) < min_genes:
                     raise Exception(f'Only {len(intersect)} HVGs were found in the intersection.\n'
                                     f'This is fewer than {min_genes} HVGs set as the minimum.\n'
                                     'Consider raising `n_stop` or reducing `min_genes`.')
                 break
-            n_hvg=int(n_hvg+step_size)
+            n_hvg = int(n_hvg + step_size)
 
     if adataOut:
-        return adata[:,list(intersect)].copy()
+        return adata[:, list(intersect)].copy()
 
     return list(intersect)
 
@@ -328,41 +328,41 @@ def hvg_batch(adata, batch_key=None, target_genes=2000, flavor='cell_ranger', n_
     then HVGs in all but one batches are used to fill up. This is continued 
     until HVGs in a single batch are considered.
     """
-    
+
     checkAdata(adata)
     if batch_key is not None:
         checkBatch(batch_key, adata.obs)
-    
+
     adata_hvg = adata if adataOut else adata.copy()
 
     n_batches = len(adata_hvg.obs[batch_key].cat.categories)
 
     # Calculate double target genes per dataset
     sc.pp.highly_variable_genes(adata_hvg,
-                                flavor=flavor, 
+                                flavor=flavor,
                                 n_top_genes=target_genes,
-                                n_bins=n_bins, 
+                                n_bins=n_bins,
                                 batch_key=batch_key)
 
     nbatch1_dispersions = adata_hvg.var['dispersions_norm'][adata_hvg.var.highly_variable_nbatches >
-                                                           len(adata_hvg.obs[batch_key].cat.categories)-1]
-    
+                                                            len(adata_hvg.obs[batch_key].cat.categories) - 1]
+
     nbatch1_dispersions.sort_values(ascending=False, inplace=True)
 
     if len(nbatch1_dispersions) > target_genes:
         hvg = nbatch1_dispersions.index[:target_genes]
-    
+
     else:
         enough = False
         print(f'Using {len(nbatch1_dispersions)} HVGs from full intersect set')
         hvg = nbatch1_dispersions.index[:]
         not_n_batches = 1
-        
+
         while not enough:
             target_genes_diff = target_genes - len(hvg)
 
             tmp_dispersions = adata_hvg.var['dispersions_norm'][adata_hvg.var.highly_variable_nbatches ==
-                                                                (n_batches-not_n_batches)]
+                                                                (n_batches - not_n_batches)]
 
             if len(tmp_dispersions) < target_genes_diff:
                 print(f'Using {len(tmp_dispersions)} HVGs from n_batch-{not_n_batches} set')
@@ -373,7 +373,7 @@ def hvg_batch(adata, batch_key=None, target_genes=2000, flavor='cell_ranger', n_
                 print(f'Using {target_genes_diff} HVGs from n_batch-{not_n_batches} set')
                 tmp_dispersions.sort_values(ascending=False, inplace=True)
                 hvg = hvg.append(tmp_dispersions.index[:target_genes_diff])
-                enough=True
+                enough = True
 
     print(f'Using {len(hvg)} HVGs')
 
@@ -381,14 +381,14 @@ def hvg_batch(adata, batch_key=None, target_genes=2000, flavor='cell_ranger', n_
         del adata_hvg
         return hvg.tolist()
     else:
-        return adata_hvg[:,hvg].copy()
+        return adata_hvg[:, hvg].copy()
 
 
 ### Feature Reduction
 def reduce_data(adata, batch_key=None, subset=False,
                 filter=True, flavor='cell_ranger', n_top_genes=2000, n_bins=20,
                 pca=True, pca_comps=50, overwrite_hvg=True,
-                neighbors=True, use_rep='X_pca', 
+                neighbors=True, use_rep='X_pca',
                 umap=True):
     """
     overwrite_hvg:
@@ -397,20 +397,20 @@ def reduce_data(adata, batch_key=None, subset=False,
         if False, skips HVG computation even if `n_top_genes` is specified and uses
         pre-existing HVG column for PCA
     """
-    
+
     checkAdata(adata)
     if batch_key:
         checkBatch(batch_key, adata.obs)
-    
+
     if n_top_genes is not None and overwrite_hvg:
         print("HVG")
-        
+
         overwrite_hvg = False
-        
+
         ## quick fix: HVG doesn't work on dense matrix
         if not sparse.issparse(adata.X):
             adata.X = sparse.csr_matrix(adata.X)
-            
+
         if batch_key is not None:
             hvg_list = hvg_batch(adata, batch_key=batch_key, target_genes=n_top_genes, n_bins=n_bins)
             adata.var['highly_variable'] = np.in1d(adata.var_names, hvg_list)
@@ -424,25 +424,25 @@ def reduce_data(adata, batch_key=None, subset=False,
 
         n_hvg = np.sum(adata.var["highly_variable"])
         print(f'Computed {n_hvg} highly variable genes')
-        
+
     if pca:
         print("PCA")
         use_hvgs = not overwrite_hvg and "highly_variable" in adata.var
         sc.tl.pca(adata,
-                  n_comps=pca_comps, 
-                  use_highly_variable=use_hvgs, 
-                  svd_solver='arpack', 
+                  n_comps=pca_comps,
+                  use_highly_variable=use_hvgs,
+                  svd_solver='arpack',
                   return_info=True)
-    
+
     if neighbors:
         print("Nearest Neigbours")
         sc.pp.neighbors(adata, use_rep=use_rep)
-    
+
     if umap:
         print("UMAP")
         sc.tl.umap(adata)
-    
-        
+
+
 ### Cell Cycle
 def score_cell_cycle(adata, organism='mouse'):
     """
@@ -454,7 +454,7 @@ def score_cell_cycle(adata, organism='mouse'):
     """
     import pathlib
     root = pathlib.Path(__file__).parent
-    
+
     cc_files = {'mouse': [root / 'resources/s_genes_tirosh.txt',
                           root / 'resources/g2m_genes_tirosh.txt'],
                 'human': [root / 'resources/s_genes_tirosh_hm.txt',
@@ -466,13 +466,13 @@ def score_cell_cycle(adata, organism='mouse'):
         g2m_genes = [x.strip() for x in f.readlines() if x.strip() in adata.var.index]
 
     if (len(s_genes) == 0) or (len(g2m_genes) == 0):
-        rand_choice = np.random.randint(1,adata.n_vars,10)
+        rand_choice = np.random.randint(1, adata.n_vars, 10)
         rand_genes = adata.var_names[rand_choice].tolist()
         raise ValueError(f"cell cycle genes not in adata\n organism: {organism}\n varnames: {rand_genes}")
-    
+
     sc.tl.score_genes_cell_cycle(adata, s_genes, g2m_genes)
 
-    
+
 def saveSeurat(adata, path, batch, hvgs=None):
     import re
     ro.r('library(Seurat)')
@@ -489,24 +489,23 @@ def saveSeurat(adata, path, batch, hvgs=None):
                 adata.layers[key].sort_indices()
 
     ro.globalenv['adata'] = adata
-    
+
     ro.r('sobj = as.Seurat(adata, counts="counts", data = "X")')
 
     # Fix error if levels are 0 and 1
     # ro.r(f'sobj$batch <- as.character(sobj${batch})')
     ro.r(f'Idents(sobj) = "{batch}"')
-    ro.r(f'saveRDS(sobj, file="{path}")') 
+    ro.r(f'saveRDS(sobj, file="{path}")')
     if hvgs is not None:
-        hvg_out = re.sub('\.RDS$', '', path)+'_hvg.RDS'
-        #hvg_out = path+'_hvg.rds'
-        ro.globalenv['hvgs']=hvgs
+        hvg_out = re.sub('\.RDS$', '', path) + '_hvg.RDS'
+        # hvg_out = path+'_hvg.rds'
+        ro.globalenv['hvgs'] = hvgs
         ro.r('unlist(hvgs)')
         ro.r(f'saveRDS(hvgs, file="{hvg_out}")')
 
-
     anndata2ri.deactivate()
-    
-       
+
+
 def readSeurat(path):
     anndata2ri.activate()
     ro.r('library(Seurat)')
@@ -515,28 +514,29 @@ def readSeurat(path):
     adata = ro.r('as.SingleCellExperiment(sobj)')
     anndata2ri.deactivate()
 
-    #Test for 'X_EMB'
+    # Test for 'X_EMB'
     if 'X_EMB' in adata.obsm:
         if 'X_emb' in adata.obsm:
             print('overwriting existing `adata.obsm["X_emb"] in the adata object')
         adata.obsm['X_emb'] = adata.obsm['X_EMB']
         del adata.obsm['X_EMB']
-    
-    return(adata)
-    
+
+    return (adata)
+
+
 def readConos(inPath):
     from time import time
     from shutil import rmtree
     from scipy.io import mmread
     from os import mkdir, path
     import pandas as pd
-    
-    dir_path = "/localscratch/conos"+str(int(time()))
+
+    dir_path = "/localscratch/conos" + str(int(time()))
     while path.isdir(dir_path):
         dir_path += '2'
     dir_path += '/'
     mkdir(dir_path)
-    
+
     ro.r('library(conos)')
     ro.r(f'con <- readRDS("{inPath}")')
     ro.r('meta <- function(sobj) {return(sobj@meta.data)}')
@@ -553,14 +553,13 @@ def readConos(inPath):
     embedding_df = pd.read_csv(dir_path + "embedding.csv")
     # Decide between using PCA or pseudo-PCA
     pseudopca_df = pd.read_csv(dir_path + "pseudopca.csv")
-    #pca_df = pd.read_csv(dir_path + "pca.csv")
+    # pca_df = pd.read_csv(dir_path + "pca.csv")
 
     graph_conn_mtx = mmread(dir_path + "graph_connectivities.mtx")
     graph_dist_mtx = mmread(dir_path + "graph_distances.mtx")
-    
-    adata = sc.read_mtx(dir_path+ "raw_count_matrix.mtx")
-    
-    
+
+    adata = sc.read_mtx(dir_path + "raw_count_matrix.mtx")
+
     adata.var_names = gene_df["gene"].values
     adata.obs_names = metadata.index.values
 
@@ -577,12 +576,10 @@ def readConos(inPath):
     adata.uns['neighbors'] = dict(connectivities=graph_conn_mtx.tocsr(), distances=graph_dist_mtx.tocsr())
 
     # Assign raw counts to .raw slot, load in normalised counts
-    #adata.raw = adata
-    #adata_temp = sc.read_mtx(DATA_PATH + "count_matrix.mtx")
-    #adata.X = adata_temp.X
+    # adata.raw = adata
+    # adata_temp = sc.read_mtx(DATA_PATH + "count_matrix.mtx")
+    # adata.X = adata_temp.X
 
     rmtree(dir_path)
-    
-    return adata
 
-    
+    return adata
