+++ scib/resources/g2m_genes_tirosh.txt
