+++ scib/resources/s_genes_tirosh.txt
