+++ tests/metrics/test_clisi.py
@@ -2,7 +2,7 @@ from tests.common import *
 
 
 def test_clisi_full(adata):
-    score = scIB.me.clisi_graph(
+    score = scib.me.clisi_graph(
         adata,
         batch_key='batch',
         label_key='celltype',
@@ -16,7 +16,7 @@ def test_clisi_full(adata):
 
 def test_clisi_embed(adata_neighbors):
     adata_neighbors.obsm['X_emb'] = adata_neighbors.obsm['X_pca']
-    score = scIB.me.clisi_graph(
+    score = scib.me.clisi_graph(
         adata_neighbors,
         batch_key='batch',
         label_key='celltype',
@@ -28,7 +28,7 @@ def test_clisi_embed(adata_neighbors):
 
 
 def test_clisi_knn(adata_neighbors):
-    score = scIB.me.clisi_graph(
+    score = scib.me.clisi_graph(
         adata_neighbors,
         batch_key='batch',
         label_key='celltype',
