+++ tests/metrics/test_graph_connectivity.py
@@ -2,6 +2,6 @@ from tests.common import *
 
 
 def test_graph_connectivity(adata_neighbors):
-    score = scIB.me.graph_connectivity(adata_neighbors, label_key='celltype')
+    score = scib.me.graph_connectivity(adata_neighbors, label_key='celltype')
     LOGGER.info(f"score: {score}")
     assert score == 0.9670013350457753
