+++ tests/metrics/test_ilisi.py
@@ -2,7 +2,7 @@ from tests.common import *
 
 
 def test_ilisi_full(adata):
-    score = scIB.me.ilisi_graph(
+    score = scib.me.ilisi_graph(
         adata,
         batch_key='batch',
         scale=True,
@@ -15,7 +15,7 @@ def test_ilisi_full(adata):
 
 def test_ilisi_embed(adata_neighbors):
     adata_neighbors.obsm['X_emb'] = adata_neighbors.obsm['X_pca']
-    score = scIB.me.ilisi_graph(
+    score = scib.me.ilisi_graph(
         adata_neighbors,
         batch_key='batch',
         scale=True,
@@ -26,7 +26,7 @@ def test_ilisi_embed(adata_neighbors):
 
 
 def test_ilisi_knn(adata_neighbors):
-    score = scIB.me.ilisi_graph(
+    score = scib.me.ilisi_graph(
         adata_neighbors,
         batch_key='batch',
         scale=True,
