+++ tests/metrics/test_pcr_metrics.py
@@ -2,12 +2,12 @@ from tests.common import *
 
 
 def test_pc_regression(adata):
-    scIB.me.pcr.pc_regression(adata.X, adata.obs["batch"])
+    scib.me.pcr.pc_regression(adata.X, adata.obs["batch"])
 
 
 def test_pcr_batch(adata):
     # no PCA precomputed
-    score = scIB.me.pcr_comparison(
+    score = scib.me.pcr_comparison(
         adata, adata,
         covariate='batch',
         n_comps=50,
@@ -18,14 +18,14 @@ def test_pcr_batch(adata):
 
 
 def test_pcr_batch_precomputed(adata_pca):
-    score = scIB.me.pcr_comparison(adata_pca, adata_pca, covariate='batch', scale=True)
+    score = scib.me.pcr_comparison(adata_pca, adata_pca, covariate='batch', scale=True)
     LOGGER.info(f"precomputed PCA: {score}")
     assert 0 <= score < 1e-6
 
 
 def test_pcr_batch_embedding(adata):
     # use different embedding
-    score = scIB.me.pcr_comparison(
+    score = scib.me.pcr_comparison(
         adata_pre=adata,
         adata_post=add_embed(adata, type_='full'),
         covariate='batch',
