+++ tests/metrics/test_silhouette_metrics.py
@@ -2,7 +2,7 @@ from tests.common import *
 
 
 def test_silhouette(adata_pca):
-    score = scIB.me.silhouette(
+    score = scib.me.silhouette(
         adata_pca,
         group_key='celltype',
         embed='X_pca',
@@ -13,7 +13,7 @@ def test_silhouette(adata_pca):
 
 
 def test_silhouette_batch(adata_pca):
-    score = scIB.me.silhouette_batch(
+    score = scib.me.silhouette_batch(
         adata_pca,
         batch_key='batch',
         group_key='celltype',
@@ -26,7 +26,7 @@ def test_silhouette_batch(adata_pca):
 
 
 def test_isolated_labels_silhouette(adata_pca):
-    score = scIB.me.isolated_labels(
+    score = scib.me.isolated_labels(
         adata_pca,
         label_key='celltype',
         batch_key='batch',
