+++ tests/metrics/test_trajectory.py
@@ -4,7 +4,7 @@ from tests.common import *
 def test_trajectory(adata_neighbors):
     sc.tl.dpt(adata_neighbors)
     adata_int = adata_neighbors.copy()
-    score = scIB.me.hvg_overlap(
+    score = scib.me.hvg_overlap(
         adata_int,
         adata_neighbors,
         batch='batch',
