+++ tests/preprocessing/test_clustering.py
@@ -2,7 +2,7 @@ from tests.common import *
 
 
 def test_cluster(adata_neighbors):
-    _, _, score_all, clustering = scIB.cl.opt_louvain(
+    _, _, score_all, clustering = scib.cl.opt_louvain(
         adata_neighbors,
         label_key='celltype',
         cluster_key='cluster',
